transcription/requirements.txt
boto3>=1.34.0
python-dotenv>=1.0.0
rich>=13.0.0
runpod>=1.6.0
